fix(ui): AG73.1 — move demo hooks into DemoOrdersView (QA hooks violations)
- Extract demo logic with conditional hooks into separate DemoOrdersView component
- Fix react-hooks/rules-of-hooks violations (hooks were called conditionally)
- Simplify AG61/AG67 demo blocks (no hooks, safe conditional rendering)
- Add Link import (next/link hygiene)

Fixes Quality Assurance failures from AG72/AG73.
